improve the dav module with the pre_request method
git-svn-id: file:///home/code/svn/ovd/trunk@1293 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ ApplicationServer/daemon/modules_fs.d/dav.sh
@@ -29,37 +29,73 @@ dav_set_fs() {
 
     [ -f ${SESSID_DIR}/parameters/module_fs/dav_dirs ] || return 1
 
-    [ -f ${SESSID_DIR}/parameters/module_fs/login ]    || return 1
-    [ -f ${SESSID_DIR}/parameters/module_fs/password ] || return 1
-    DAV_LOGIN=$(cat ${SESSID_DIR}/parameters/module_fs/login)
-    DAV_PASSWORD=$(cat ${SESSID_DIR}/parameters/module_fs/password)
-    check_variables DAV_LOGIN DAV_PASSWORD || return 1
+    if [ -f ${SESSID_DIR}/parameters/module_fs/cookie_url ]; then
+	DAV_AUTH=url
+	DAV_INIT_URL=$(cat ${SESSID_DIR}/parameters/module_fs/cookie_url)
 
+	check_variables DAV_INIT_URL || return 1
+
+    elif [ -f ${SESSID_DIR}/parameters/module_fs/login ] && \
+	[ -f ${SESSID_DIR}/parameters/module_fs/password ]; then
+	DAV_AUTH=password
+	DAV_LOGIN=$(cat ${SESSID_DIR}/parameters/module_fs/login)
+	DAV_PASSWORD=$(cat ${SESSID_DIR}/parameters/module_fs/password)
+	check_variables DAV_LOGIN DAV_PASSWORD || return 1
+    else
+	log_ERROR "dav: missing authentication method"
+	return 1
+    fi
+
+    DAV_CONFIG=${SESSID_DIR}/private/davfs2.conf
+    DAV_SECRET=${SESSID_DIR}/private/davfs2.secret
     DAV_MOUNT_POINT=/mnt/dav/$USER_LOGIN
-    DAV_CREDENTIALS=/etc/davfs2/secrets
 }
 
 dav_get_status() {
-    local_get_status || return 1    
+    local_get_status && return 0
 
-    local dirs=$(find $DAV_MOUNT_POINT -type d -maxdepth 1 -mindepth 1)
-    for dir in $dirs; do
-    is_mount_point  $dir || return 1
-    done
+    [ -d $DAV_MOUNT_POINT ] && return 0
+
+    return 1
+}
+
+dav_do_init_config() {
+    echo "ask_auth 0"      >$DAV_CONFIG
+
+    if [ $DAV_AUTH == 'url' ]; then
+	echo "allow_cookie 1" >>$DAV_CONFIG
+	echo "pre_request $DAV_INIT_URL" >>$DAV_CONFIG
+    else
+	echo "secrets $DAV_SECRET">>$DAV_CONFIG
+
+	while read line; do
+            local url=$(echo $line| cut -d '|' -f2)
+	    echo "$url $DAV_LOGIN $DAV_PASSWORD" >>$DAV_SECRET
+	done < ${SESSID_DIR}/parameters/module_fs/dav_dirs
+	chmod 600 $DAV_SECRET
+    fi
+}
+
+dav_do_purge_config() {
+    rm $DAV_CONFIG
+    [ -f $DAV_SECRET ] && rm $DAV_SECRET
 }
 
 dav_do_mount() {
     local_do_mount || return 1
+    dav_do_init_config || return 1
 
     local i=0
     while read line; do
         local name=$(echo $line| cut -d '|' -f1)
         local url=$(echo $line| cut -d '|' -f2)
 
-        echo "$url $DAV_LOGIN $DAV_PASSWORD" >>$DAV_CREDENTIALS
-
         mkdir -p $DAV_MOUNT_POINT/$i
-        mount -t davfs -o uid=$USER_ID,dir_mode=700,file_mode=600 $url $DAV_MOUNT_POINT/$i
+        mount -t davfs -o conf=$DAV_CONFIG,uid=$USER_ID,dir_mode=700,file_mode=600 "$url" $DAV_MOUNT_POINT/$i
+        if [ $? != 0 ]; then
+            log_ERROR "dav: mount failed from '$url'"
+            return 1
+        fi
 
         mkdir -p "$USER_HOME/$name"
         mount --bind $DAV_MOUNT_POINT/$i  "$USER_HOME/$name"
@@ -72,8 +108,9 @@ dav_do_mount() {
 }
 
 dav_do_umount() {
-    dav_do_umount_bind && dav_do_umount_real
-
+    dav_do_purge_config
+    dav_do_umount_bind
+    dav_do_umount_real
     rmdir $DAV_MOUNT_POINT
 
     local_do_umount
@@ -81,18 +118,19 @@ dav_do_umount() {
 
 
 dav_do_umount_real() {
-    local i=0
-    while read line; do
-        local name=$(echo $line| cut -d '|' -f1)
-        local url=$(echo $line| cut -d '|' -f2)
+    local nb=$(wc -l ${SESSID_DIR}/parameters/module_fs/dav_dirs |cut  -d ' ' -f1)
 
+    for i in $(seq 0 $nb); do
         local dir=$DAV_MOUNT_POINT/$i
-        umount "$dir" || log_ERROR "dav: Failed to umount $dir"
-        sed -i "\%^$url %d" $DAV_CREDENTIALS
-        rmdir $DAV_MOUNT_POINT/$i
+        is_mount_point "$dir" || continue
 
-        i=$(( $i + 1 ))
-    done < ${SESSID_DIR}/parameters/module_fs/dav_dirs
+        umount "$dir" || log_ERROR "dav: Failed to umount $dir"
+    done
+
+    for i in $(seq 0 $nb); do
+        [ -d $DAV_MOUNT_POINT/$i ] || continue
+        rmdir $DAV_MOUNT_POINT/$i
+    done
 }
 
 
@@ -102,6 +140,8 @@ dav_do_umount_bind() {
         local url=$(echo $line| cut -d '|' -f2)
 
         local dir="$USER_HOME/$name"
+        is_mount_point "$dir" || continue
+
         log_INFO "dav: umounting bind $dir"
         umount "$dir" || log_ERROR "dav: Failed to umount $dir"
 
@@ -112,6 +152,7 @@ dav_do_umount_bind() {
 dav_do_umount_critical() {
     local dirs=`find $DAV_MOUNT_POINT -maxdepth 1 -mindepth 1 -type d`
     for dir in $dirs; do
+	is_mount_point "$dir" || continue
 	umount "$dir"
 	rmdir "$dir"
     done
